# java-quarkus/generated/swagger-ui/quarkus_swagger.yaml.yaml
...
swagger: "2.0"
info:
  description: "jump the queue implementation in Java Quarkus"
  version: "v4"
  title: "JTQ in Quarkus"
schemes:
- "https"
paths:
...
  /queuemanagement/v1/queue/{id}:
    get:
      operationId: "getQueue"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/QueueEto"
    delete:
      operationId: "deleteQueue"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "integer"
            format: "int64"
...
definitions:
...
  QueueEto:
    type: "object"
    properties:
      modificationCounter:
        type: "integer"
        format: "int32"
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
      logo:
        type: "string"
      currentNumber:
        type: "string"
      attentionTime:
        type: "string"
        format: "date-time"
      minAttentionTime:
        type: "string"
        format: "date-time"
      active:
        type: "boolean"
...